Fixed clang warnings

+++ source/newtonian/test_2d/random_pert.cpp
@@ -6,11 +6,11 @@ vector<Vector2D> add_random_pert(vector<Vector2D> const& original,
 {
   srand(0);
   vector<Vector2D> res = original;
-  for(int i=0;i<(int)res.size();++i){
+  for(size_t i=0;i<res.size();++i){
     res[i].x += 
-      amp_x*2*((double)rand()/(double)RAND_MAX-0.5);
+      amp_x*2*(static_cast<double>(rand())/static_cast<double>(RAND_MAX)-0.5);
     res[i].y += 
-      amp_y*2*((double)rand()/(double)RAND_MAX-0.5);
+      amp_y*2*(static_cast<double>(rand())/static_cast<double>(RAND_MAX)-0.5);
   }
   return res;
 }
